CL-643: Generalize more DAO business logic by introducing low level database API in DAOs: SchemaDao adaptation + get rid of deprecated transformToRest in entities
