+++ databases/orientdb/src/main/java/com/syncleus/ferma/ext/orientdb3/TxDataImpl.java
@@ -5,6 +5,8 @@ import com.gentics.mesh.cli.OrientDBBootstrapInitializer;
 import com.gentics.mesh.core.data.HibMeshVersion;
 import com.gentics.mesh.core.data.dao.PermissionRoots;
 import com.gentics.mesh.core.db.TxData;
+import com.gentics.mesh.core.link.WebRootLinkReplacer;
+import com.gentics.mesh.core.search.index.node.NodeIndexHandler;
 import com.gentics.mesh.etc.config.AuthenticationOptions;
 import com.gentics.mesh.etc.config.CacheConfig;
 import com.gentics.mesh.etc.config.ClusterOptions;
@@ -20,6 +22,8 @@ import com.gentics.mesh.etc.config.OrientDBMeshOptions;
 import com.gentics.mesh.etc.config.VertxOptions;
 import com.gentics.mesh.etc.config.search.ElasticSearchOptions;
 
+import io.vertx.core.Vertx;
+
 /**
  * @see TxData
  */
@@ -28,11 +32,17 @@ public class TxDataImpl implements TxData {
 	private final OrientDBBootstrapInitializer boot;
 	private final OrientDBMeshOptions options;
 	private final PermissionRoots permissionRoots;
+	private final NodeIndexHandler nodeIndexHandler;
+	private final WebRootLinkReplacer webRootLinkReplacer;
 
-	public TxDataImpl(OrientDBMeshOptions options, OrientDBBootstrapInitializer boot, PermissionRoots permissionRoots) {
+	public TxDataImpl(OrientDBMeshOptions options, OrientDBBootstrapInitializer boot, 
+			PermissionRoots permissionRoots, NodeIndexHandler nodeIndexHandler,
+			WebRootLinkReplacer webRootLinkReplacer) {
 		this.options = options;
 		this.boot = boot;
 		this.permissionRoots = permissionRoots;
+		this.nodeIndexHandler = nodeIndexHandler;
+		this.webRootLinkReplacer = webRootLinkReplacer;
 	}
 
 	@Override
@@ -333,6 +343,21 @@ public class TxDataImpl implements TxData {
 	@Override
 	public PermissionRoots permissionRoots() {
 		return permissionRoots;
+	}
+
+	@Override
+	public Vertx vertx() {
+		return boot.vertx();
+	}
+
+	@Override
+	public NodeIndexHandler nodeIndexHandler() {
+		return nodeIndexHandler;
+	}
+
+	@Override
+	public WebRootLinkReplacer webRootLinkReplacer() {
+		return webRootLinkReplacer;
 	}
 
 }
