+++ mdm/api/src/main/java/com/gentics/mesh/core/data/dao/JobDao.java
@@ -1,6 +1,7 @@
 package com.gentics.mesh.core.data.dao;
 
 import java.time.ZonedDateTime;
+import java.util.Map;
 import java.util.function.Predicate;
 
 import com.gentics.mesh.context.BulkActionContext;
@@ -10,10 +11,13 @@ import com.gentics.mesh.core.data.job.HibJob;
 import com.gentics.mesh.core.data.page.Page;
 import com.gentics.mesh.core.data.perm.InternalPermission;
 import com.gentics.mesh.core.data.project.HibProject;
+import com.gentics.mesh.core.data.schema.HibMicroschema;
 import com.gentics.mesh.core.data.schema.HibMicroschemaVersion;
+import com.gentics.mesh.core.data.schema.HibSchema;
 import com.gentics.mesh.core.data.schema.HibSchemaVersion;
 import com.gentics.mesh.core.data.user.HibUser;
 import com.gentics.mesh.core.rest.job.JobResponse;
+import com.gentics.mesh.core.rest.job.JobWarningList;
 import com.gentics.mesh.event.EventQueueBatch;
 import com.gentics.mesh.parameter.PagingParameters;
 
@@ -184,4 +188,61 @@ public interface JobDao extends DaoGlobal<HibJob>, DaoTransformable<HibJob, JobR
 	 * Delete all jobs.
 	 */
 	void clear();
+
+	@Override
+	default JobResponse transformToRestSync(HibJob job, InternalActionContext ac, int level, String... languageTags) {
+		JobResponse response = new JobResponse();
+		response.setUuid(job.getUuid());
+
+		HibUser creator = job.getCreator();
+		if (creator != null) {
+			response.setCreator(creator.transformToReference());
+		} else {
+			//log.error("The object {" + getClass().getSimpleName() + "} with uuid {" + getUuid() + "} has no creator. Omitting creator field");
+		}
+
+		String date = job.getCreationDate();
+		response.setCreated(date);
+		response.setErrorMessage(job.getErrorMessage());
+		response.setErrorDetail(job.getErrorDetail());
+		response.setType(job.getType());
+		response.setStatus(job.getStatus());
+		response.setStopDate(job.getStopDate());
+		response.setStartDate(job.getStartDate());
+		response.setCompletionCount(job.getCompletionCount());
+		response.setNodeName(job.getNodeName());
+
+		JobWarningList warnings = job.getWarnings();
+		if (warnings != null) {
+			response.setWarnings(warnings.getData());
+		}
+
+		Map<String, String> props = response.getProperties();
+		HibBranch branch = job.getBranch();
+		if (branch != null) {
+			props.put("branchName", branch.getName());
+			props.put("branchUuid", branch.getUuid());
+		} else {
+			log.debug("No referenced branch found.");
+		}
+
+		HibSchemaVersion toSchema = job.getToSchemaVersion();
+		if (toSchema != null) {
+			HibSchema container = toSchema.getSchemaContainer();
+			props.put("schemaName", container.getName());
+			props.put("schemaUuid", container.getUuid());
+			props.put("fromVersion", job.getFromSchemaVersion().getVersion());
+			props.put("toVersion", toSchema.getVersion());
+		}
+
+		HibMicroschemaVersion toMicroschema = job.getToMicroschemaVersion();
+		if (toMicroschema != null) {
+			HibMicroschema container = toMicroschema.getSchemaContainer();
+			props.put("microschemaName", container.getName());
+			props.put("microschemaUuid", container.getUuid());
+			props.put("fromVersion", job.getFromMicroschemaVersion().getVersion());
+			props.put("toVersion", toMicroschema.getVersion());
+		}
+		return response;
+	}
 }
