+++ mdm/api/src/main/java/com/gentics/mesh/core/data/group/HibGroup.java
@@ -10,9 +10,7 @@ import com.gentics.mesh.core.TypeInfo;
 import com.gentics.mesh.core.data.HibBucketableElement;
 import com.gentics.mesh.core.data.HibCoreElement;
 import com.gentics.mesh.core.data.HibNamedElement;
-import com.gentics.mesh.core.data.dao.GroupDao;
 import com.gentics.mesh.core.data.user.HibUserTracking;
-import com.gentics.mesh.core.db.Tx;
 import com.gentics.mesh.core.rest.group.GroupReference;
 import com.gentics.mesh.core.rest.group.GroupResponse;
 import com.gentics.mesh.handler.VersionUtils;
@@ -59,11 +57,5 @@ public interface HibGroup extends HibCoreElement<GroupResponse>, HibUserTracking
 	@Override
 	default String getAPIPath(InternalActionContext ac) {
 		return VersionUtils.baseRoute(ac) + "/groups/" + getUuid();
-	}
-
-	@Override
-	default GroupResponse transformToRestSync(InternalActionContext ac, int level, String... languageTags) {
-		GroupDao groupDao = Tx.get().groupDao();
-		return groupDao.transformToRestSync(this, ac, level, languageTags);
 	}
 }
