+++ mdm/api/src/main/java/com/gentics/mesh/core/data/node/HibMicronode.java
@@ -1,5 +1,10 @@
 package com.gentics.mesh.core.data.node;
 
+import static com.gentics.mesh.core.rest.error.Errors.error;
+import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
+
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import com.gentics.mesh.context.InternalActionContext;
@@ -9,8 +14,13 @@ import com.gentics.mesh.core.data.HibNodeFieldContainer;
 import com.gentics.mesh.core.data.HibTransformableElement;
 import com.gentics.mesh.core.data.diff.FieldContainerChange;
 import com.gentics.mesh.core.data.schema.HibMicroschemaVersion;
+import com.gentics.mesh.core.db.Tx;
 import com.gentics.mesh.core.rest.micronode.MicronodeResponse;
+import com.gentics.mesh.core.rest.node.field.Field;
+import com.gentics.mesh.core.rest.schema.FieldSchema;
+import com.gentics.mesh.core.rest.schema.MicroschemaModel;
 import com.gentics.mesh.core.result.Result;
+import com.gentics.mesh.parameter.impl.NodeParametersImpl;
 
 public interface HibMicronode extends HibFieldContainer, HibBaseElement, HibTransformableElement<MicronodeResponse> {
 
@@ -67,5 +77,40 @@ public interface HibMicronode extends HibFieldContainer, HibBaseElement, HibTran
 	default String getAPIPath(InternalActionContext ac) {
 		// Micronodes have no public location
 		return null;
+	}
+
+	default MicronodeResponse transformToRestSync(InternalActionContext ac, int level, String... languageTags) {
+
+		NodeParametersImpl parameters = new NodeParametersImpl(ac);
+		MicronodeResponse restMicronode = new MicronodeResponse();
+		HibMicroschemaVersion microschemaContainer = getSchemaContainerVersion();
+		if (microschemaContainer == null) {
+			throw error(BAD_REQUEST, "The microschema container for micronode {" + getUuid() + "} could not be found.");
+		}
+
+		MicroschemaModel microschemaModel = microschemaContainer.getSchema();
+		if (microschemaModel == null) {
+			throw error(BAD_REQUEST, "The microschema for micronode {" + getUuid() + "} could not be found.");
+		}
+
+		restMicronode.setMicroschema(microschemaContainer.transformToReference());
+		restMicronode.setUuid(getUuid());
+
+		List<String> requestedLanguageTags = new ArrayList<>();
+		if (languageTags.length == 0) {
+			requestedLanguageTags.addAll(parameters.getLanguageList(Tx.get().data().options()));
+		} else {
+			requestedLanguageTags.addAll(Arrays.asList(languageTags));
+		}
+
+		// Fields
+		for (FieldSchema fieldEntry : microschemaModel.getFields()) {
+			Field restField = getRestField(ac, fieldEntry.getName(), fieldEntry, requestedLanguageTags, level);
+			if (restField != null) {
+				restMicronode.getFields().put(fieldEntry.getName(), restField);
+			}
+		}
+
+		return restMicronode;
 	}
 }
