+++ mdm/api/src/main/java/com/gentics/mesh/core/data/schema/HibFieldSchemaVersionElement.java
@@ -6,12 +6,9 @@ import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.HibCoreElement;
 import com.gentics.mesh.core.data.branch.HibBranch;
 import com.gentics.mesh.core.data.job.HibJob;
-import com.gentics.mesh.core.data.schema.handler.FieldSchemaContainerComparator;
 import com.gentics.mesh.core.rest.schema.FieldSchemaContainer;
 import com.gentics.mesh.core.rest.schema.FieldSchemaContainerVersion;
-import com.gentics.mesh.core.rest.schema.change.impl.SchemaChangesListModel;
 import com.gentics.mesh.core.result.Result;
-import com.gentics.mesh.event.EventQueueBatch;
 import com.gentics.mesh.util.StreamUtil;
 import com.gentics.mesh.util.VersionUtil;
 
@@ -155,39 +152,6 @@ public interface HibFieldSchemaVersionElement<R extends FieldSchemaContainer, RM
 	}
 
 	/**
-	 * Generate a schema change list by comparing the schema with the specified schema update model which is extracted from the action context.
-	 *
-	 * @param ac
-	 *            Action context that provides the schema update request
-	 * @param comparator
-	 *            Comparator to be used to compare the schemas
-	 * @param restModel
-	 *            Rest model of the container that should be compared
-	 * @return Rest model which contains the changes list
-	 */
-	SchemaChangesListModel diff(InternalActionContext ac, FieldSchemaContainerComparator<?> comparator, FieldSchemaContainer restModel);
-
-	/**
-	 * Apply changes which will be extracted from the action context.
-	 *
-	 * @param ac
-	 *            Action context that provides the migration request data
-	 * @param batch
-	 * @return The created schema container version
-	 */
-	SCV applyChanges(InternalActionContext ac, EventQueueBatch batch);
-
-	/**
-	 * Apply the given list of changes to the schema container. This method will invoke the schema migration process.
-	 *
-	 * @param ac
-	 * @param listOfChanges
-	 * @param batch
-	 * @return The created schema container version
-	 */
-	SCV applyChanges(InternalActionContext ac, SchemaChangesListModel listOfChanges, EventQueueBatch batch);
-
-	/**
 	 * Get the branches to which the container was assigned.
 	 *
 	 * @return Found branches of this version
@@ -207,6 +171,20 @@ public interface HibFieldSchemaVersionElement<R extends FieldSchemaContainer, RM
 	 * @return
 	 */
 	Iterable<? extends HibJob> referencedJobsViaFrom();
+
+	/**
+	 * Transform the element into the matching rest model response synchronously.
+	 *
+	 * @param ac
+	 *            Context of the calling action
+	 * @param level
+	 *            Level of transformation
+	 * @param languageTags
+	 *            optional list of language tags to be used for language fallback
+	 * @return
+	 * @deprecated Use DAO method to transform elements instead
+	 */
+	R transformToRestSync(InternalActionContext ac, int level, String... languageTags);
 
 	default int compareTo(SCV version) {
 		return VersionUtil.compareVersions(getVersion(), version.getVersion());
