+++ mdm/api/src/main/java/com/gentics/mesh/core/data/tag/HibTag.java
@@ -6,11 +6,9 @@ import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.HibBucketableElement;
 import com.gentics.mesh.core.data.HibCoreElement;
 import com.gentics.mesh.core.data.HibNamedElement;
-import com.gentics.mesh.core.data.dao.TagDao;
 import com.gentics.mesh.core.data.project.HibProject;
 import com.gentics.mesh.core.data.tagfamily.HibTagFamily;
 import com.gentics.mesh.core.data.user.HibUserTracking;
-import com.gentics.mesh.core.db.Tx;
 import com.gentics.mesh.core.rest.tag.TagReference;
 import com.gentics.mesh.core.rest.tag.TagResponse;
 import com.gentics.mesh.handler.VersionUtils;
@@ -72,15 +70,5 @@ public interface HibTag extends HibCoreElement<TagResponse>, HibUserTracking, Hi
 	@Override
 	default String getAPIPath(InternalActionContext ac) {
 		return VersionUtils.baseRoute(ac) + "/" + encodeSegment(getProject().getName()) + "/tagFamilies/" + getTagFamily().getUuid() + "/tags/" + getUuid();
-	}
-
-	/**
-	 * Use {@link TagDaoWrapper#transformToRestSync(HibTag, InternalActionContext, int, String...)} instead.
-	 */
-	@Override
-	@Deprecated
-	default TagResponse transformToRestSync(InternalActionContext ac, int level, String... languageTags) {
-		TagDao tagDao = Tx.get().tagDao();
-		return tagDao.transformToRestSync(this, ac, level, languageTags);
 	}
 }
