+++ mdm/api/src/main/java/com/gentics/mesh/core/data/user/HibUser.java
@@ -10,9 +10,7 @@ import com.gentics.mesh.core.TypeInfo;
 import com.gentics.mesh.core.data.HibBucketableElement;
 import com.gentics.mesh.core.data.HibCoreElement;
 import com.gentics.mesh.core.data.HibNamedElement;
-import com.gentics.mesh.core.data.dao.UserDao;
 import com.gentics.mesh.core.data.node.HibNode;
-import com.gentics.mesh.core.db.Tx;
 import com.gentics.mesh.core.rest.user.UserReference;
 import com.gentics.mesh.core.rest.user.UserResponse;
 import com.gentics.mesh.handler.VersionUtils;
@@ -307,12 +305,5 @@ public interface HibUser extends HibCoreElement<UserResponse>, HibUserTracking, 
 	@Override
 	default String getAPIPath(InternalActionContext ac) {
 		return VersionUtils.baseRoute(ac) + "/users/" + getUuid();
-	}
-
-	@Deprecated
-	@Override
-	default UserResponse transformToRestSync(InternalActionContext ac, int level, String... languageTags) {
-		UserDao userDao = Tx.get().userDao();
-		return userDao.transformToRestSync(this, ac, level, languageTags);
 	}
 }
