+++ mdm/api/src/main/java/com/gentics/mesh/core/link/WebRootLinkReplacer.java
