+++ mdm/orientdb-wrapper/src/main/java/com/gentics/mesh/core/data/dao/AbstractRootDaoWrapper.java
@@ -1,6 +1,7 @@
 package com.gentics.mesh.core.data.dao;
 
 import com.gentics.mesh.cli.OrientDBBootstrapInitializer;
+import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.HibCoreElement;
 import com.gentics.mesh.core.data.MeshCoreVertex;
 import com.gentics.mesh.core.data.generic.PermissionPropertiesImpl;
@@ -21,7 +22,7 @@ import dagger.Lazy;
  * @param <R> MDM API root entity type 
  */
 public abstract class AbstractRootDaoWrapper<RM extends RestModel, L extends HibCoreElement<RM>, D extends MeshCoreVertex<RM>, R extends HibCoreElement<? extends RestModel>> 
-	extends AbstractDaoWrapper<L> {
+	extends AbstractDaoWrapper<L> implements DaoTransformable<L, RM> {
 
 	public AbstractRootDaoWrapper(Lazy<OrientDBBootstrapInitializer> boot, Lazy<PermissionPropertiesImpl> permissions) {
 		super(boot, permissions);
@@ -29,7 +30,7 @@ public abstract class AbstractRootDaoWrapper<RM extends RestModel, L extends Hib
 
 	@SuppressWarnings("unchecked")
 	public L persist(R root, String uuid) {
-		D vertex = getRoot(root).createRaw();
+		D vertex = getRoot(root).create();
 		L entity = (L) vertex;
 		if (uuid != null) {
 			entity.setUuid(uuid);
@@ -40,6 +41,11 @@ public abstract class AbstractRootDaoWrapper<RM extends RestModel, L extends Hib
 
 	public void unpersist(R root, L element) {
 		getRoot(root).findByUuid(element.getUuid()).remove();
+	}
+
+	@Override
+	public String getETag(L element, InternalActionContext ac) {
+		return element.getETag(ac);
 	}
 
 	/**
