+++ mdm/orientdb-wrapper/src/main/java/com/gentics/mesh/core/data/dao/impl/LanguageDaoWrapperImpl.java
@@ -189,4 +189,15 @@ public class LanguageDaoWrapperImpl extends AbstractCoreDaoWrapper<LanguageRespo
 	protected RootVertex<Language> getRoot() {
 		return boot.get().meshRoot().getLanguageRoot();
 	}
+
+	@Override
+	public LanguageResponse transformToRestSync(HibLanguage element, InternalActionContext ac, int level,
+			String... languageTags) {
+		LanguageResponse model = new LanguageResponse();
+		model.setUuid(element.getUuid());
+		model.setLanguageTag(element.getLanguageTag());
+		model.setName(element.getName());
+		model.setNativeName(element.getNativeName());
+		return model;
+	}
 }
