+++ mdm/orientdb-wrapper/src/main/java/com/gentics/mesh/core/data/dao/impl/ProjectDaoWrapperImpl.java
@@ -314,7 +314,6 @@ public class ProjectDaoWrapperImpl extends AbstractCoreDaoWrapper<ProjectRespons
 		setRolePermissions(graphProject, ac, restProject);
 
 		return restProject;
-
 	}
 
 	@Override
