+++ mdm/orientdb-wrapper/src/main/java/com/gentics/mesh/core/data/dao/impl/TagFamilyDaoWrapperImpl.java
@@ -121,7 +121,6 @@ public class TagFamilyDaoWrapperImpl extends AbstractCoreDaoWrapper<TagFamilyRes
 		}
 
 		return restTagFamily;
-
 	}
 
 	@Override
